remove encoding
