+++ lib/s3-client.js
@@ -1,7 +1,7 @@
 const AWS = require('aws-sdk');
 const constants = require('./constants');
 const validator = require('./s3-validator-replacer');
-const { binaryDecode, binaryEncode } = require('./binaryEncoding');
+
 
 class S3Client {
     constructor() {
@@ -24,14 +24,6 @@ class S3Client {
             ...options
         };
         this._s3 = new AWS.S3(opt);
-        if (options.binary) {
-            this._parse = binaryDecode;
-            this._stringify = binaryEncode;
-        }
-        else {
-            this._parse = JSON.parse;
-            this._stringify = JSON.stringify;
-        }
     }
 
     /**
@@ -43,7 +35,7 @@ class S3Client {
      * @api public
      */
     async put(options) {
-        const params = validator.validateAndReplacePutParams(options, this._stringify);
+        const params = validator.validateAndReplacePutParams(options);
         return this._s3.upload(params).promise();
     }
 
@@ -69,7 +61,7 @@ class S3Client {
     async get(options) {
         const params = validator.validateAndReplaceGetParams(options);
         const data = await this._s3.getObject(params).promise();
-        return this._parse(data.Body);
+        return data.Body;
     }
 
     /**
