+++ lib/s3-validator-replacer.js
@@ -14,13 +14,13 @@ class S3ValidatorReplacer {
         return { ...params, Bucket: this._parseBucket(params.Bucket) };
     }
 
-    validateAndReplacePutParams(params, encode = JSON.stringify) {
+    validateAndReplacePutParams(params) {
         this._validateSchema(this._putSchema, params);
         return {
             ...params,
             Bucket: this._parseBucket(params.Bucket),
             Key: this._parseKey(params.Key),
-            Body: this._parseBody(params.Body, encode)
+            Body: this._parseBody(params.Body)
         };
     }
 
@@ -57,11 +57,11 @@ class S3ValidatorReplacer {
         return key;
     }
 
-    _parseBody(body, encode = JSON.stringify) {
+    _parseBody(body) {
         if (body instanceof stream || body instanceof Buffer) {
             return body;
         }
-        return encode(body);
+        return body;
     }
 }
