+++ package.json
@@ -15,10 +15,10 @@
   "license": "ISC",
   "dependencies": {
     "ajv": "^6.6.2",
-    "bson": "^4.0.3",
     "aws-sdk": "^2.382.0"
   },
   "devDependencies": {
+    "@hkube/encoding": "^1.0.7",
     "chai": "^4.2.0",
     "coveralls": "^3.0.2",
     "eslint": "^4.19.1",
